Stabilize deadlines tests
