+++ jbpm-container-test/jbpm-in-container-test/jbpm-container-test-suite/src/test/java/org/jbpm/test/container/archive/EJBService.java
@@ -41,7 +41,7 @@ import org.jbpm.test.container.JbpmContainerTest;
 import org.jbpm.test.container.listeners.TrackingAgendaEventListener;
 import org.jbpm.test.container.tools.IntegrationMavenResolver;
 import org.jbpm.test.container.webspherefix.WebSphereFixedJtaPlatform;
-import org.jbpm.test.listener.DefaultCountDownProcessEventListener;
+import org.jbpm.test.listener.process.DefaultCountDownProcessEventListener;
 import org.kie.api.KieServices;
 import org.kie.api.builder.KieBuilder;
 import org.kie.api.builder.KieFileSystem;
