+++ jbpm-services/jbpm-services-ejb/jbpm-services-ejb-impl/src/test/java/org/jbpm/services/ejb/test/AsyncExecutionProcessServiceEJBIntegrationTest.java
@@ -36,7 +36,7 @@ import org.jbpm.services.api.model.DeploymentUnit;
 import org.jbpm.services.ejb.api.DeploymentServiceEJBLocal;
 import org.jbpm.services.ejb.api.ProcessServiceEJBLocal;
 import org.jbpm.services.ejb.api.RuntimeDataServiceEJBLocal;
-import org.jbpm.test.listener.NodeLeftCountDownProcessEventListener;
+import org.jbpm.test.listener.process.NodeLeftCountDownProcessEventListener;
 import org.junit.After;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -64,7 +64,7 @@ public class AsyncExecutionProcessServiceEJBIntegrationTest extends AbstractTest
 		}
 		WebArchive war = ShrinkWrap.createFromZipFile(WebArchive.class, archive);
 		war.addPackage("org.jbpm.services.ejb.test"); // test cases
-		war.addClass("org.jbpm.test.listener.NodeLeftCountDownProcessEventListener");
+		war.addClass("org.jbpm.test.listener.process.NodeLeftCountDownProcessEventListener");
 		// deploy test kjar
 		deployKjar();
