+++ jbpm-test-util/src/main/java/org/jbpm/test/listener/process/DefaultCountDownProcessEventListener.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.jbpm.test.listener;
+package org.jbpm.test.listener.process;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
